* More test fixes.
